Namespace the fileset indexer to CHF
