+++ app/models/file_set.rb
@@ -3,7 +3,7 @@ class FileSet < ActiveFedora::Base
   include ::CurationConcerns::FileSetBehavior
   include Sufia::FileSetBehavior
 
-  self.indexer = ::FileSetIndexer
+  self.indexer = CHF::FileSetIndexer
 
   def create_derivatives(filename)
     # use layer 0 to create our derivatives
