+++ spec/models/file_set_spec.rb
@@ -2,6 +2,6 @@ require 'rails_helper'
 
 RSpec.describe FileSet do
   it 'uses custom indexer' do
-    expect(FileSet.indexer).to eq FileSetIndexer
+    expect(FileSet.indexer).to eq CHF::FileSetIndexer
   end
 end
